Update viewport columns to handle colspan
