+++ src/hooks/useViewportRows.ts
@@ -50,9 +50,9 @@ export function useViewportRows<R>({
     return groupRows(rawRows, groupBy, 0);
   }, [groupBy, rowGrouper, rawRows]);
 
-  const [rows, allGroupRows] = useMemo(() => {
+  const [rows, isGroupRow] = useMemo(() => {
     const allGroupRows = new Set<unknown>();
-    if (!groupedRows) return [rawRows, allGroupRows];
+    if (!groupedRows) return [rawRows, isGroupRow];
 
     const flattenedRows: Array<R | GroupRow<R>> = [];
     const expandGroup = (rows: GroupByDictionary<R> | readonly R[], parentId: string | undefined, level: number): void => {
@@ -87,10 +87,13 @@ export function useViewportRows<R>({
     };
 
     expandGroup(groupedRows, undefined, 0);
-    return [flattenedRows, allGroupRows];
+    return [flattenedRows, isGroupRow];
+
+    function isGroupRow(row: R | GroupRow<R>): row is GroupRow<R> {
+      return allGroupRows.has(row);
+    }
   }, [expandedGroupIds, groupedRows, rawRows]);
 
-  const isGroupRow = <R>(row: unknown): row is GroupRow<R> => allGroupRows.has(row);
 
   if (!enableVirtualization) {
     return {
